requirements.txt
numpy==1.23.4
pandas==1.5.1
python-dateutil==2.8.2
python-dotenv==0.21.0
loguru==0.6.0
